introduce ratequery.hpp
